requirements.txt
streamlit
langchain
pypdf
faiss-cpu
sentence-transformers
textblob